// projects/kl-calculate/src/lib/models/billing.dto.ts
import {BillingLineDto} from './billing-line.dto';
import {BillingDiscountDto} from './billing-discount.dto';

export class BillingDto {
  id?: number;
  external_id?: string;
  company?: any;
  working_period?: any;
  type?: string;
  document_type?: string;
  document_status?: string;
  guid?: string;
  document_num?: string;
  document_date?: number;
  external_document_num?: string;
  branch?: any;
  current_account?: any;
  warehouse?: any;
  currency?: any;
  exchange_rate?: number;
  sales_person?: any;
  payment_plan?: any;
  shipment_address?: any;
  carrier_company?: any;
  special_code?: string;
  expense_amount?: number;
  discount_amount?: number;
  vat_amount?: number;
  without_vat_amount?: number;
  net_total_amount?: number;
  gross_total_amount?: number;
  rec_status?: string;
  lines?: Array<BillingLineDto>;
  discounts?: Array<BillingDiscountDto>;
  erp_transferred?: string;
  erp_document_num?: string;
  explanation?: string;
  estate_document_num?: string;
  gib_status_code?: string;
  gib_status_description?: string;
  gib_date?: any;
  accept_denied_status?: string;
  integrator_status_description?: string;
  integrator_status_code?: number;
  accept_denied_description?: string;
  document_tracking_type?: any;
  document_tracking_id?: number;
  document_tracking_num?: string;
}